feat(generator): add support for React including HMR and unit testing
