+++ README.md
@@ -42,7 +42,7 @@
 - Unit testing 
 - System testing (browser-based)
 - Support for releasing code using semantic releasing, conventional commits
-- Supports Angular 1.x and Angular 2.x frameworks, and more coming (PRs welcome!!!) 
+- Supports Angular 1.x, Angular 2.x (beta) and now React 15.x (PRs welcome!!!) 
 - Sample projects for some configurations (see below)
 - EXTENSIBLE! Decorate / extend configurations then re-run `yo confit` to re-generate configuration while preserving your changes (uses tags to identify generated-areas inside files) 
 
@@ -86,7 +86,7 @@ Confit is a Yeoman generator that generates web-development tools for the main d
 The tooling is generated by answering a series of simple questions about your project, turning a process that used to take weeks to
 tune correctly, into a 5 minute step. See the animated-gif above for an example.
  
-***Out of the box, Confit can generate a sample project for the settings you've chosen and so you can see that everything works!***
+***Out of the box, Confit can generate a sample project for the settings you've chosen so that you can see that everything works!***
 
 ### In more detail...
 
@@ -114,7 +114,7 @@ There are sample project implementations for the following technology combinatio
 
 Build Profile | JS Source Format | JS Frameworks Supported (optional)
 :------------ | :--------------- | :----------------------
-Browser       | ES6              | Angular 1.x, Angular 2
+Browser       | ES6              | Angular 1.x, Angular 2, React
 Browser       | Typescript       | Angular 1.x, Angular 2
 Node          | ES6              |
