+++ lib/buildTool/webpack/sampleApp/templates/react/demoModule/DemoService.js
@@ -1,6 +1,8 @@
 export default {
   demo: function(param) {
-    window.console.log('Message from demoService.demo():' + param);
+    let msg = 'Message from demoService.demo():' + param;
+    window.console.log(msg);
+    return msg;
   }
 };
