+++ lib/buildTool/webpack/sampleApp/templates/react/demoModule/unitTest/unitTest.spec.js
@@ -18,4 +18,8 @@ describe('Test imported module', function () {
   it('should have a demoService with a demo method', function () {
     expect(typeof demoService.demo).toEqual('function');
   });
+
+  it('should return a message when demo() is called', function () {
+    expect(demoService.demo('abc')).toEqual('Message from demoService.demo():abc');
+  });
 });
